chat_information_nl_to_sql_bi_agent_orchestrator_with_memory_first_try/backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
#sqlalchemy==2.1.0
sqlalchemy
pyodbc==4.0.35
pandas==2.2.2
langchain_openai==0.1.8
#langchain_core==0.1.22
langchain_core
python-multipart==0.0.6
qdrant-client
openai
#openai==0.28
tqdm
python-dotenv
sentence-transformers 
requests
torch

